feat(features) Introduce Custom Exception handling. Implement Fluent Validation before creating 'Leave Request'

+++ HR.LeaveManagement.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using HR.LeaveManagement.Application.Contracts.Persistence;
+using HR.LeaveManagement.Application.Exceptions;
 using MediatR;
 using System.Drawing;
 
@@ -19,6 +20,8 @@ public class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeComm
         var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
 
         // verify 
+        if (leaveTypeToDelete == null)
+            throw new NotFoundException(nameof(leaveTypeToDelete), request.Id);
 
         // Integrate the DB 
         await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
